+++ libs/sage-monorepo/nx-sage-monorepo/src/plugins/build-project-targets.ts
@@ -1,0 +1,16 @@
+import { TargetConfiguration } from '@nx/devkit';
+import { buildImageTarget } from './build-image-target';
+import { SageMonorepoProjectConfiguration } from './sage-monorepo-project-configuration';
+import { SageMonorepoPluginConfiguration } from './sage-monorepo-plugin-configuration';
+
+export async function buildProjectTargets(
+  projectRoot: string,
+  config: SageMonorepoPluginConfiguration,
+): Promise<SageMonorepoProjectConfiguration> {
+  const targets: Record<string, TargetConfiguration> = {};
+
+  targets[config.buildImageTargetName] = await buildImageTarget(projectRoot);
+
+  const metadata = {};
+  return { targets, metadata };
+}
