+++ libs/sage-monorepo/nx-sage-monorepo/src/plugins/sage-monorepo-plugin-configuration.ts
@@ -1,0 +1,13 @@
+import { SageMonorepoPluginOptions } from './sage-monorepo-plugin-options';
+
+export interface SageMonorepoPluginConfiguration {
+  buildImageTargetName: string;
+}
+
+export function createPluginConfiguration({
+  buildImageTargetName = 'build-image',
+}: SageMonorepoPluginOptions): SageMonorepoPluginConfiguration {
+  return {
+    buildImageTargetName,
+  };
+}
