+++ libs/sage-monorepo/nx-sage-monorepo/src/plugins/sage-monorepo-plugin-options.ts
@@ -1,0 +1,3 @@
+export interface SageMonorepoPluginOptions {
+  buildImageTargetName?: string;
+}
