+++ libs/sage-monorepo/nx-sage-monorepo/src/plugins/sage-monorepo-project-configuration.ts
@@ -1,0 +1,3 @@
+import { ProjectConfiguration } from '@nx/devkit';
+
+export type SageMonorepoProjectConfiguration = Pick<ProjectConfiguration, 'targets' | 'metadata'>;
